// src/openapi/models/IndexProductResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * imperia-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { PaginationMeta } from "./PaginationMeta";
import {
  PaginationMetaFromJSON,
  PaginationMetaFromJSONTyped,
  PaginationMetaToJSON,
} from "./PaginationMeta";
import type { Product } from "./Product";
import {
  ProductFromJSON,
  ProductFromJSONTyped,
  ProductToJSON,
} from "./Product";

/**
 * Index products response object.
 * @export
 * @interface IndexProductResponse
 */
export interface IndexProductResponse {
  /**
   *
   * @type {Array<Product>}
   * @memberof IndexProductResponse
   */
  data: Array<Product>;
  /**
   *
   * @type {PaginationMeta}
   * @memberof IndexProductResponse
   */
  meta: PaginationMeta;
  /**
   *
   * @type {string}
   * @memberof IndexProductResponse
   */
  message: string;
}

/**
 * Check if a given object implements the IndexProductResponse interface.
 */
export function instanceOfIndexProductResponse(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "data" in value;
  isInstance = isInstance && "meta" in value;
  isInstance = isInstance && "message" in value;

  return isInstance;
}

export function IndexProductResponseFromJSON(json: any): IndexProductResponse {
  return IndexProductResponseFromJSONTyped(json, false);
}

export function IndexProductResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): IndexProductResponse {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    data: (json["data"] as Array<any>).map(ProductFromJSON),
    meta: PaginationMetaFromJSON(json["meta"]),
    message: json["message"],
  };
}

export function IndexProductResponseToJSON(
  value?: IndexProductResponse | null
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    data: (value.data as Array<any>).map(ProductToJSON),
    meta: PaginationMetaToJSON(value.meta),
    message: value.message,
  };
}