// src/openapi/models/IndexServiceResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * imperia-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { PaginationMeta } from "./PaginationMeta";
import {
  PaginationMetaFromJSON,
  PaginationMetaFromJSONTyped,
  PaginationMetaToJSON,
} from "./PaginationMeta";
import type { Service } from "./Service";
import {
  ServiceFromJSON,
  ServiceFromJSONTyped,
  ServiceToJSON,
} from "./Service";

/**
 * Index services response object.
 * @export
 * @interface IndexServiceResponse
 */
export interface IndexServiceResponse {
  /**
   *
   * @type {Array<Service>}
   * @memberof IndexServiceResponse
   */
  data: Array<Service>;
  /**
   *
   * @type {PaginationMeta}
   * @memberof IndexServiceResponse
   */
  meta: PaginationMeta;
  /**
   *
   * @type {string}
   * @memberof IndexServiceResponse
   */
  message: string;
}

/**
 * Check if a given object implements the IndexServiceResponse interface.
 */
export function instanceOfIndexServiceResponse(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "data" in value;
  isInstance = isInstance && "meta" in value;
  isInstance = isInstance && "message" in value;

  return isInstance;
}

export function IndexServiceResponseFromJSON(json: any): IndexServiceResponse {
  return IndexServiceResponseFromJSONTyped(json, false);
}

export function IndexServiceResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): IndexServiceResponse {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    data: (json["data"] as Array<any>).map(ServiceFromJSON),
    meta: PaginationMetaFromJSON(json["meta"]),
    message: json["message"],
  };
}

export function IndexServiceResponseToJSON(
  value?: IndexServiceResponse | null
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    data: (value.data as Array<any>).map(ServiceToJSON),
    meta: PaginationMetaToJSON(value.meta),
    message: value.message,
  };
}